refactor(AuthService.kt): add custom exceptions for 404 and 401 errors
- UnauthorizedException.kt class
- NotFoundException.kt class
- GlobalExceptionHandler.kt handles new types of exceptions
- AuthService.kt throws appropriate exceptions and status codes
- 401 on incorrect password
- 404 on login with non-registered account
- 201 instead of 200 on member registration in AuthService.kt
