+++ src/main/kotlin/ecommerce/exception/GlobalExceptionHandler.kt
@@ -1,5 +1,6 @@
-package ecommerce.advice
+package ecommerce.exception
 
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ControllerAdvice
@@ -13,9 +14,18 @@ class GlobalExceptionHandler {
         return ResponseEntity.badRequest().body(errors)
     }
 
-    // TODO maybe split function into handling sameNameException and other exceptions to be less generic
     @ExceptionHandler(IllegalArgumentException::class)
     fun handleIllegalArgument(e: IllegalArgumentException): ResponseEntity<String> {
         return ResponseEntity.badRequest().body(e.message)
+    }
+
+    @ExceptionHandler(NotFoundException::class)
+    fun handleNotFound(e: NotFoundException): ResponseEntity<String> {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.message)
+    }
+
+    @ExceptionHandler(UnauthorizedException::class)
+    fun handleUnauthorized(e: UnauthorizedException): ResponseEntity<String> {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.message)
     }
 }
