+++ src/main/kotlin/ecommerce/exception/NotFoundException.kt
@@ -1,0 +1,8 @@
+package ecommerce.exception
+
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.bind.annotation.ResponseStatus
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+class NotFoundException(override val message: String?) : RuntimeException()
