+++ src/main/kotlin/ecommerce/exception/UnauthorizedException.kt
@@ -1,0 +1,7 @@
+package ecommerce.exception
+
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ResponseStatus
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+class UnauthorizedException(override val message: String?) : RuntimeException()
